requirements.txt
blinker==1.6.2
click==8.1.3
colorama==0.4.6
Flask==2.3.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
flask-marshmallow==0.15.0
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
mysqlclient==2.1.1
packaging==23.1
PyJWT==2.6.0
PyMySQL==1.0.3
six==1.16.0
SQLAlchemy==2.0.11
typing_extensions==4.5.0
Werkzeug==2.3.2
